admin_panel/requirements.txt
Flask
redis
python-dotenv
python-telegram-bot
apscheduler
